research_pipeline/requirements-research.txt
# Extra deps for research_pipeline/ (install atop models/requirements.txt)
sentence-transformers>=2.2.0
trl>=0.7.0
peft>=0.10.0
accelerate>=0.25.0
datasets>=2.14.0
scipy>=1.11.0
